Added BlockCache which saves an amount of block id -> address associations in a smart way
Shortcuts of blocks, which aren't taken often (written into BlockCache) will be removed first if space is low in BlockCache. So, each block is assigned  a generation, which will be pushed on write. If there is no space left, the block with the lowest generation will be removed from cache. If there are multiple blocks, that share the same lowest generation, which is removed is not determined (because of underlying STL implementation of unordered maps).

Current limitation:
- A block id cannot be 0, since 0 is used as a read failure indicator