Fixed invalid memory access when using a foreign buffer in protocol messages
By assigning a reference to a member variable, the object the reference points to will be deleted when the scope of this member variable ends. Because Buffer mData is a class member variable the Buffer reference passed into
the constructor will be deleted when the class' scope ends.
By using a reference as a member variable this problem should dismiss.