Added SSL session resumption
SSL session ids/tickets are saved on a per-thread basis. This means,
each thread has a limited cache.
For efficient caching functionality BlockCache is used.
Sessions are associated with a host and port and will be used
automatically if there is an existing session. This means, there can
only exist one SSL session for each host-port combination (subdomains
count as separate hosts).