Added toInt and toEnum functions
These functions convert an enum to an int or an int to an enum,
respectively. Invalid enum values will be mapped to a new auto generated
enum value "INVALID_ENUM_VALUE".

To realize this, I moved common generator functions to common.py. This
resulted in a necessary file structure change.